TeNumericFieldProperty
